build: FATAL_ASSERTS flag, useful for qemu debug
ML ASSERT() is a very misleading name.  It's included
in all builds and is non-fatal.  Consequently, if you see
the message in qemu, execution has proceeded past the
point of the error and your state is all confused.

FATAL_ASSERTS=y at build time replaces the existing
handler with while(1).  This can make inspecting state
that triggered an assert much easier.