mlv_lite: check for division by zero
This tries to make the mlv_lite code, and associated
core code, safe re zeroes in division.  I wouldn't be surprised
if I've missed some.

It's expected this doesn't change behaviour on D45 cams.
It is required on D6 cams as those trigger an exception.
And in any case, we shouldn't trigger undefined behaviour in C.