modules: add hiding via .hid files in MODULES/
This allows hiding modules from the GUI.
They can still load, either by manually creating a .en file,
or if the auto-loader uses them to solve a dependency when
another module is enabled.

This both allows hiding modules that *can load but shouldn't*,
and modules that can load but don't want to be visible.

The former case allows us to make modules safer on new cams,
that might have all the deps met, but the module code hard-codes
MMIO that are untested, for example.

The latter allows implementing modules that act like libraries;
silently loaded on demand, no need for a GUI as they provide
nothing on their own.