200D: mostly fix 10 and 12 bit raw video modes
These now record 10 and 12 bit data.  But every other frame
is a stuck image, which is sometimes junk data.  14 bit reliably
gets a good image on all frames.

I don't understand why this occurs.  I suspect it's a problem
in our code, since if you film a stopwatch, you can see that the
*first* recorded frame may not be the earliest time on the stopwatch.
We're recording out of order.